# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: AI Agent Marketplace API
  version: 1.0.0
  description: API documentation for the AI agent marketplace
  contact:
    name: API Support
    email: support@agenx.com
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://api.agenx.com
    description: Production server
...
paths:
...
  /agents/{agentId}/reviews:
    get:
      tags:
        - Reviews
      summary: Get reviews for a specific agent
      description: Retrieve a paginated list of reviews for a specific agent with optional filtering and sorting.
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, highest, lowest]
            default: newest
          description: Sort order for reviews
        - name: rating
          in: query
          schema:
            type: integer
            enum: [1, 2, 3, 4, 5]
          description: Filter reviews by rating
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Reviews
      summary: Submit a review for a specific agent
      description: Create a new review for an agent. Users can only submit one review per agent.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
      responses:
        '201':
          description: Review submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  review:
                    $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Invalid input or duplicate review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agentId}/reviews/summary:
    get:
      tags:
        - Reviews
      summary: Get review summary for a specific agent
      description: Retrieve a summary of reviews for a specific agent including average rating and distribution.
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
      responses:
        '200':
          description: Review summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  summary:
                    $ref: '#/components/schemas/ReviewSummary'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews/{reviewId}:
    put:
      tags:
        - Reviews
      summary: Edit a review
      description: Edit an existing review. Reviews can only be edited within 48 hours of submission.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the review
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  review:
                    $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not owner or edit window expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Reviews
      summary: Delete a review
      description: Delete an existing review. Users can only delete their own reviews.
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the review
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Review deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateReviewRequest:
      type: object
      required:
        - rating
        - content
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Rating from 1 to 5 stars
        content:
          type: string
          minLength: 10
          maxLength: 2000
          description: Review content
        images:
          type: array
          items:
            type: string
            format: uri
            description: Image URL or Base64 encoded image
          maxItems: 5
          description: Optional images to include with the review
    UpdateReviewRequest:
      type: object
      required:
        - rating
        - content
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Rating from 1 to 5 stars
        content:
          type: string
          minLength: 10
          maxLength: 2000
          description: Review content
        images:
          type: array
          items:
            type: string
            format: uri
            description: Image URL or Base64 encoded image
          maxItems: 5
          description: Optional images to include with the review
...
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            avatar:
              type: string
              nullable: true
            isVerified:
              type: boolean
            isCurrentUser:
              type: boolean
            isOfficial:
              type: boolean
              nullable: true
        rating:
          type: integer
        date:
          type: string
          format: date-time
        formattedDate:
          type: string
        content:
          type: string
        replies:
          type: array
          items:
            $ref: '#/components/schemas/ReviewReplyResponse'
        replyCount:
          type: integer
        helpful:
          type: object
          properties:
            upvotes:
              type: integer
            downvotes:
              type: integer
            userVote:
              type: integer
        additionalImages:
          type: array
          items:
            $ref: '#/components/schemas/ReviewImageResponse'
    ReviewReplyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/ReviewAuthor'
        date:
          type: string
          format: date-time
        formattedDate:
          type: string
        content:
          type: string
    ReviewImageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        thumbnailUrl:
          type: string
        alt:
          type: string
          nullable: true
    ReviewAuthor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatar:
          type: string
          nullable: true
        isVerified:
          type: boolean
        isCurrentUser:
          type: boolean
        isOfficial:
          type: boolean
          nullable: true
    ReviewSummary:
      type: object
      properties:
        averageRating:
          type: number
          format: float
        totalReviews:
          type: integer
        credibilityScore:
          type: number
          format: float
        credibilityBadge:
          type: string
          enum: [excellent, good, average, poor, not-rated]
        recentPositivePercentage:
          type: integer
        ratingDistribution:
          type: object
          properties:
            '1':
              type: integer
            '2':
              type: integer
            '3':
              type: integer
            '4':
              type: integer
            '5':
              type: integer

    # Response wrappers
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string